change bedtools to samtools for coverage calculation. no re-sorting. --resume addition

+++ bin/paqman.sh
@@ -30,6 +30,7 @@ meryldb=""
 merylmem="10"
 merylkmer="21"
 localbuscodb=""
+resume="no"
 help="nohelp"
 
 ## to clean up a bunch of output from the tools in order to reduce all the unnecessary output
@@ -125,6 +126,10 @@ case "$key" in
 	shift
 	shift
 	;;
+	--resume)
+    resume="yes"
+	shift
+	;;
 	-c|--cleanup)
 	cleanup="$2"
 	shift
@@ -154,10 +159,11 @@ case "$key" in
 	-p | --prefix       Prefix for output (default: name of assembly file (-a) before the fasta suffix)
 	-o | --output       Name of output folder for all results (default: paqman_output)
 	-seq | --sequences	Whether or not to use scaffolds or contigs; provide 'scaffolds' to not break the assembly at N's (default: contigs)
-	-mdb | --meryldb	A precomputed Meryl database for your dataset. Generated automatically if not provided.
-	-mm | --merylmem	The soft RAM limit in GB used whilst building the Meryl database (default: 10)
-	-mk | --merylkmer	The k-mer size used to build the Meryl database (default: 21)
+	-mdb | --meryldb	A precomputed meryl database for your dataset. Generated automatically if not provided.
+	-mm | --merylmem	The soft RAM limit in GB used whilst building the meryl database (default: 10)
+	-mk | --merylkmer	The k-mer size used to build the meryl database (default: 21)
 	-lbdb | --localbuscodb	A predownloaded busco database for your dataset. Downloaded automatically if not provided.
+	--resume			Resume a incomplete run of PAQman. Incomplete steps will be rerun from scratch.
 	-c | --cleanup      Remove a large number of files produced by each of the tools that can take up a lot of space. Choose between 'yes' or 'no' (default: yes)
 	-h | --help         Print this help message
 	"
@@ -229,7 +235,14 @@ fi
 [[ $platform != "ont" && $platform != "pacbio-hifi" && $platform != "pacbio-clr" ]] && echo "ERROR: --platform option needs to be 'ont' (default) or 'pacbio-hifi' or 'pacbio-clr'" && exit
 [[ $cleanup != "yes" && $cleanup != "no" ]] && echo "ERROR: --cleanup option neither 'yes' (default) or 'no'" && exit
 
-[ -d "${output}" ] && echo "ERROR: output folder already exists" && exit
+[[ -d "${output}" && $resume == "no" ]] && echo "ERROR: output folder already exists (use --resume to restart run from last step)" && exit
+[[ ! -d "$output" && "$resume" == "yes" ]] && echo "ERROR: --resume was set but output folder doesn't exist" && exit
+
+## the window size for median-average binning (in bp)
+window2=$( echo $window | awk '{print $0/1000}' )
+## distance for the window to move before recalculating the window (in bp)
+slide2=$( echo $slide | awk '{print $0/1000}' )
+
 
 ##############################################################
 #################### BEGINNING EVALUATION ####################
@@ -237,10 +250,15 @@ fi
 echo "#################################################"
 echo "################ Starting PAQman ################"
 echo "#################################################"
+
+########################## SET UP ##########################
+##check if step is to be run based on stream variable
+##begin step 1
 echo "$(date +%H:%M) ########## Step 1: Organising Input"
 
 ## assembly evaluations
-mkdir ${output}
+
+[[ ${resume} == "no"  ]] && mkdir ${output}
 cd ${output}
 
 ##organise the assemblies
@@ -300,6 +318,8 @@ assembly="${assembly2}.fa"
 fi
 fi
 
+##index the assembly for a couple downstream steps
+samtools faidx ${assembly}
 
 ## make symbolic links to the input files in the output folder
 ln -sf ${longreadpath} ./
@@ -311,16 +331,62 @@ pair12=$( echo ${pair1} | awk -F "/" '{print $NF}' )
 pair22=$( echo ${pair2} | awk -F "/" '{print $NF}' )
 
 
+########################## RESUME CHECK ##########################
+stream="step1,step2,step3,step4,step5,step6,step7"
+
+##check if resume parameter has been set then find those steps that were not completed (based on the file complete.tmp in the respective folder)
+if [[ ${resume} == "yes"  ]]
+then
+stream="step1"  # always start with step1
+
+##define the steps in order
+folders=(quast busco merqury craq coverage telomerality)
+
+##define the corresponding step numbers
+stepnums=(2 3 4 5 6 7)
+
+##Loop over folders
+for i in "${!folders[@]}"
+do
+folder="${folders[$i]}"
+step="step${stepnums[$i]}"
+
+##Check if folder exists and complete.tmp is missing
+if [[ ! -f "${folder}/complete.tmp" ]]; then
+##Append to stream variable
+stream="${stream},${step}"
+fi
+done
+
+stream2=$( echo $stream | sed 's/step1/Step1:Set-up; /g' | sed 's/,step2/Step2:Quast; /g' | sed 's/,step3/Step3:BUSCO; /g' | sed 's/,step4/Step4:Merqury; /g' | sed 's/,step5/Step5:CRAQ; /g' | sed 's/,step6/Step6:Coverage; /g' | sed 's/,step7/Step7:Telomerality/g' )
+
+echo "Resuming steps: $stream2"
+fi
+
 
 ########################## QUAST ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step2,"* ]]; then
+[ -e "./quast" ] && rm -r ./quast
+
+##begin step 2
 echo "$(date +%H:%M) ########## Step 2: Running Quast"
+
 quast -o ./quast ${assembly} > quast.log
 ##move quast log to quast output folder
 mv quast.log quast/
 ## we are just interested in the summary tsv file 'quast/report.tsv'
 
+touch ./quast/complete.tmp
+##close stream variable if check
+fi
 
 ########################## BUSCO (using the ${buscodb} dataset) ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step3,"* ]]; then
+[ -e "./busco" ] && rm -r ./busco
+
+##begin step 3
 echo "$(date +%H:%M) ########## Step 3: Running BUSCO"
 
 if [[ $localbuscodb != "" ]]
@@ -345,8 +411,19 @@ fi
 ## we are just interested in the summary txt file 'busco/short_summary.*.txt'
 
 
+touch ./busco/complete.tmp
+##close stream variable if check
+fi
+
 ########################## Merqury ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step4,"* ]]; then
+[ -e "./merqury" ] && rm -r ./merqury
+
+##begin step 4
 echo "$(date +%H:%M) ########## Step 4a: Generating k-mer distribution"
+
+
 if [[ $meryldb != "" ]]
 then
 echo "NOTE: Using precomputed meryl k-mer database" 
@@ -368,7 +445,7 @@ fi
 
 fi
 
-echo "$(date +%H:%M) ########## Step 5b: Running Merqury"
+echo "$(date +%H:%M) ########## Step 4b: Running Merqury"
 
 mkdir ./merqury
 if [[ $meryldb != "" ]]
@@ -399,9 +476,18 @@ rm -r *.meryl
 fi
 
 
+touch ./merqury/complete.tmp
+##close stream variable if check
+fi
 
 ########################## CRAQ ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step5,"* ]]; then
+[ -e "./craq" ] && rm -r ./craq
+
+##begin step 5
 echo "$(date +%H:%M) ########## Step 5a: Downsampling for 50X long-reads for CRAQ assessment"
+
 ## redownsample the dataset for just 50X of the longest as to remove shorter reads from confusing CRAQ
 ## get genome size based on input genome
 genomesize=$( cat ./quast/report.tsv  | awk -F "\t" '{if(NR == 15) print $2}' )
@@ -437,25 +523,28 @@ fi
 ## from the output we are just interested in the summary file 'craq/runAQI_out/out_final.Report' for the summary stats (second line from the top is the genome wide average)
 ## for the position of structural errors, 'craq/runAQI_out/strER_out/out_final.CSE.bed'
 
-
-
+touch ./craq/complete.tmp
+##close stream variable if check
+fi
 
 ########################## READ COVERAGE ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step6,"* ]]; then
+[ -e "./coverage" ] && rm -r ./coverage
+
+##begin step 6
 echo "$(date +%H:%M) ########## Step 6a: Running whole genome read alignment"
+
 ## using the short-read data we can look at genome wide coverage on the assembly
 ## first index the assembly using the aligner
 mkdir ./coverage
 bwa index ${assembly}
 
-## the window size for median-average binning (in bp)
-window2=$( echo $window | awk '{print $0/1000}' )
-## distance for the window to move before recalculating the window (in bp)
-slide2=$( echo $slide | awk '{print $0/1000}' )
+
 ## create a bed file use for binning
-samtools faidx ${assembly}
 cat ${assembly}.fai | cut -f1-2 > ${assembly}.bed
 ## now split the reference bed file up into the above prescribed bins
-bedtools makewindows -w ${window} -s ${slide} -g ${assembly}.bed > ./coverage/${prefix}.${window2}kbwindow_${slide2}kbslide.bed
+bedtools makewindows -w ${window} -s ${slide} -g ${assembly}.bed  > ./coverage/${prefix}.${window2}kbwindow_${slide2}kbslide.bed
 ## get the coverage file (slightly modify by giving a range for the single basepair coverage value) then use bedtools map to overlap with the reference derived window file to create median-averaged bins
 
 ###RUNNING THE LONG-READ ALIGNMENT
@@ -464,7 +553,7 @@ bedtools makewindows -w ${window} -s ${slide} -g ${assembly}.bed > ./coverage/${
 [[ $platform == "pacbio-clr" ]] && minimap2 --secondary=no -ax map-pb -t ${threads} ${assembly} longreads.rasusa.fq.gz | samtools sort -@ 4 -o ./coverage/${prefix}.minimap.sorted.bam -
 
 ## get the coverage
-bedtools genomecov -d -split -ibam ./coverage/${prefix}.minimap.sorted.bam | gzip > ./coverage/${prefix}.minimap.sorted.cov.tsv.gz
+samtools depth -a -d 0 -@ 4 ./coverage/${prefix}.minimap.sorted.bam  | gzip > ./coverage/${prefix}.minimap.sorted.cov.tsv.gz
 
 ## calculate the median across the whole genome using the exact basepair
 medianLR=$( zcat ./coverage/${prefix}.minimap.sorted.cov.tsv.gz | awk '{if($3 != "0") print $3}' | sort -n | awk '{ a[i++]=$1} END{x=int((i+1)/2); if(x < (i+1)/2) print (a[x-1]+a[x])/2; else print a[x-1];}' )
@@ -498,7 +587,7 @@ then
 bwa mem -t ${threads} ${assembly} ${pair12} ${pair22} | samtools sort -@ 4 -o ./coverage/${prefix}.bwamem.sorted.bam
 
 ## get the coverage
-bedtools genomecov -d -split -ibam ./coverage/${prefix}.bwamem.sorted.bam | gzip > ./coverage/${prefix}.bwamem.sorted.cov.tsv.gz
+samtools depth -a -d 0 -@ 4 ./coverage/${prefix}.bwamem.sorted.bam | gzip > ./coverage/${prefix}.bwamem.sorted.cov.tsv.gz
 
 ## calculate the median across the whole genome using the exact basepair
 medianSR=$( zcat ./coverage/${prefix}.bwamem.sorted.cov.tsv.gz | awk '{if($3 != "0") print $3}' | sort -n | awk '{ a[i++]=$1} END{x=int((i+1)/2); if(x < (i+1)/2) print (a[x-1]+a[x])/2; else print a[x-1];}' )
@@ -543,9 +632,18 @@ Rscript ./coverage/${prefix}.coverage_plots.R
 fi
 
 
+touch ./coverage/complete.tmp
+##close stream variable if check
+fi
 
 ########################## TELOMERALITY ##########################
+##check if step is to be run based on stream variable
+if [[ ",$stream," == *",step7,"* ]]; then
+[ -e "./telomerality" ] && rm -r ./telomerality
+
+##begin step 7
 echo "$(date +%H:%M) ########## Step 7: Running Telomere search"
+
 mkdir ./telomerality
 #telomererepeat="TTAGGG"
 ##can also label all regions with the canonical telomeric repeat
@@ -573,17 +671,23 @@ cat ./telomerality/telomeres.bed | awk -v contig="$contig" '{if($1 == contig) pr
 cat ./telomerality/telomeres.bed | awk -v contig="$contig" '{if($1 == contig) print}' | tail -n1 | awk -v contig="$contig" -v size="$size" '{if($1 ==contig && (size-$3) <= 0.75*($3-$2)) {print contig"\tend\t"(size-$3)"\t"$2"-"$3"\ttelomeric"; end="noneed"} else if($1 ==contig && (size-$3) > 0.75*($3-$2) && (size-$3) < 5000) {print contig"\tend\t"(size-$3)"\t"$2"-"$3"\tdistant"; end="noneed"}} END{if(end != "noneed") print contig"\tend\tNA\tNA\tabsent"}'
 done >> ./telomerality/telomeres.classification.tsv
 
+touch ./telomerality/complete.tmp
+##close stream variable if check
+fi
 
 ########################## LAST CLEAN-UP ##########################
 ## just removing some index files, links etc
 if [[ $cleanup == "yes" ]]
 then
-rm ${assembly}.fai
+[ -e "${assembly}.fai" ] &&  rm ${assembly}.fai
 fi
+
 
 ########################## SUMMARY STATS ##########################
 
+##begin step 8
 echo "$(date +%H:%M) ########## Step 8: Generating summary statistics file"
+
 ### Create the header for the summary stats file
 if [[ ${sequences} == "scaffolds" ]]
 then
@@ -620,6 +724,7 @@ phredval=$( cat ./merqury/${prefix}.merqury.qv | awk '{print $4}' )
 if [[ $phredval == "+inf" ]]
 then
 echo "WARNING: Merqury detected no errors; QV will be estimated using the rule of three"
+genomesize=$( cat ./quast/report.tsv  | awk -F "\t" '{if(NR == 15) print $2}' )
 phredval=$( echo "scale=5; 10 * (l($genomesize)/l(10) - l(3)/l(10))" | bc -l )
 fi
 ##combine both
@@ -653,6 +758,10 @@ telomeralitystat=$(  echo "${telomericends};${t2t}" | tr ';' '\t' )
 ##spit out all the stats and save them to the summary stats file
 echo "${prefix};${assembly2};${quaststat};${buscostat};${merqurystat};${craqstat};${covstat};${telomeralitystat}" | tr ';' '\t' >> summary_stats.tsv
 
+
+##remove complete.tmp files
+rm */complete.tmp
+
 echo "$(date +%H:%M) ######### Summary stats can be found here ${output}/summary_stats.tsv"
-echo "################ All complete; thanks for using PAQman"
+echo "################ All complete; thanks for using PAQman; E noho rā"
